# openapi.yaml
components:
  schemas:
    AccessRequestCreate:
      properties:
        email:
          format: email
          title: Email
          type: string
        full_name:
          title: Full Name
          type: string
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
      required:
      - full_name
      - email
      title: AccessRequestCreate
      type: object
    AccessRequestResponse:
      properties:
        created_at:
          title: Created At
          type: string
        email:
          title: Email
          type: string
        full_name:
          title: Full Name
          type: string
        id:
          title: Id
          type: integer
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
        rejection_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Rejection Reason
        reviewed_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Reviewed At
        reviewed_by:
          anyOf:
          - type: integer
          - type: 'null'
          title: Reviewed By
        status:
          title: Status
          type: string
        updated_at:
          title: Updated At
          type: string
      required:
      - id
      - full_name
      - email
      - reason
      - status
      - reviewed_by
      - reviewed_at
      - rejection_reason
      - created_at
      - updated_at
      title: AccessRequestResponse
      type: object
    AccessRequestUpdate:
      properties:
        rejection_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Rejection Reason
        status:
          title: Status
          type: string
      required:
      - status
      title: AccessRequestUpdate
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
  securitySchemes:
    OAuth2PasswordBearer:
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
      type: oauth2
info:
  description: Backend API for cybercrime investigation dashboard
  title: Cybercrime Investigation API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /api/v1/access-requests/request:
    post:
      description: Submit a new investigator access request
      operationId: create_access_request_api_v1_access_requests_request_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessRequestCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessRequestResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Access Request
      tags:
      - access-requests
  /api/v1/access-requests/requests:
    get:
      description: Get all access requests (superadmin only)
      operationId: get_access_requests_api_v1_access_requests_requests_get
      parameters:
      - in: query
        name: status_filter
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status Filter
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AccessRequestResponse'
                title: Response Get Access Requests Api V1 Access Requests Requests
                  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Access Requests
      tags:
      - access-requests
  /api/v1/access-requests/requests/{request_id}:
    get:
      description: Get a specific access request
      operationId: get_access_request_api_v1_access_requests_requests__request_id__get
      parameters:
      - in: path
        name: request_id
        required: true
        schema:
          title: Request Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessRequestResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Access Request
      tags:
      - access-requests
  /api/v1/access-requests/requests/{request_id}/review:
    patch:
      description: Approve or reject an access request (superadmin only)
      operationId: review_access_request_api_v1_access_requests_requests__request_id__review_patch
      parameters:
      - in: path
        name: request_id
        required: true
        schema:
          title: Request Id
          type: integer
      - in: query
        name: reviewed_by
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Reviewed By
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessRequestUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessRequestResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Review Access Request
      tags:
      - access-requests
...